// packages/datadog-api-client-v2/models/SecurityMonitoringRuleCase.d.ts
/**
 * Unless explicitly stated otherwise all files in this repository are licensed under the Apache-2.0 License.
 * This product includes software developed at Datadog (https://www.datadoghq.com/).
 * Copyright 2020-Present Datadog, Inc.
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { SecurityMonitoringRuleSeverity } from './SecurityMonitoringRuleSeverity';
/**
* Case when signal is generated.
*/
export declare class SecurityMonitoringRuleCase {
    /**
    * A rule case contains logical operations (`>`,`>=`, `&&`, `||`) to determine if a signal should be generated based on the event counts in the previously defined queries.
    */
    'condition'?: string;
    /**
    * Name of the case.
    */
    'name'?: string;
    /**
    * Notification targets for each rule case.
    */
    'notifications'?: Array<string>;
    'status'?: SecurityMonitoringRuleSeverity;
    static readonly discriminator: string | undefined;
    static readonly attributeTypeMap: {
        [key: string]: {
            baseName: string;
            type: string;
            format: string;
        };
    };
    static getAttributeTypeMap(): {
        [key: string]: {
            baseName: string;
            type: string;
            format: string;
        };
    };
    static deserialize(data: {
        [key: string]: any;
    }): SecurityMonitoringRuleCase;
    static serialize(data: SecurityMonitoringRuleCase): {
        [key: string]: any;
    };
    constructor();
}